% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity_check.R
\name{linearity_check}
\alias{linearity_check}
\title{Linear regression model plot with extracted R-squared and p-value}
\usage{
linearity_check(x, y, ...)
}
\arguments{
\item{x}{numeric vector for independent variable}

\item{y}{numeric vector for dependent variable}

\item{...}{used to check that only x and y are included as inputs. Not needed as input variable.}
}
\value{
numeric value of adjusted R-squared and p-value, and the plot
}
\description{
Linear regression model plot with extracted R-squared and p-value
}
\details{
The function plots two variables x versus y and applies a linear regression model to calculate the adjusted R-squared and p-value
}
\examples{
x <- c(0:9)
y <- c(5,6,11,13,18,17,21,23,26,27)
linearity_check(x,y)

if (requireNamespace("datateachr", quietly = TRUE)) {
  cancer_sample <- datateachr::cancer_sample
  x <- cancer_sample$radius_mean
  y <- cancer_sample$perimeter_mean
  linearity_check(x,y)
}

}
